+++ server/ItemsService/ItemsService.Service/Interfaces/IItemService.cs
@@ -7,6 +7,6 @@ namespace ItemsService.Service.Interfaces
     public interface IItemService
     {
         public IEnumerable<Item> GetMaxPrices();
-        public double GetMaxPriceForItem(string itemName);
+        public decimal GetMaxPriceForItem(string itemName);
     }
 }
